+++ src/crested/utils/__init__.py
@@ -1,6 +1,10 @@
 """General utility functions used across CREsted."""
 
 from ._logging import setup_logging
+from ._old import (
+    EnhancerOptimizer,
+    derive_intermediate_sequences,
+)
 from ._seq_utils import (
     flip_region_strand,
     hot_encoding_to_sequence,
@@ -9,9 +13,7 @@ from ._seq_utils import (
     reverse_complement,
 )
 from ._utils import (
-    EnhancerOptimizer,
     calculate_nucleotide_distribution,
-    derive_intermediate_sequences,
     fetch_sequences,
     read_bigwig_region,
 )
