+++ src/crested/utils/_utils.py
@@ -5,8 +5,6 @@ from __future__ import annotations
 import os
 import random
 import re
-from collections.abc import Callable
-from typing import Any
 
 import numpy as np
 from anndata import AnnData
@@ -14,80 +12,6 @@ from anndata import AnnData
 from crested._genome import Genome, _resolve_genome
 from crested._io import _extract_tracks_from_bigwig
 from crested.utils._seq_utils import one_hot_encode_sequence
-
-
-class EnhancerOptimizer:
-    """
-    Class to optimize the mutated sequence based on the original prediction.
-
-    Can be passed as the 'enhancer_optimizer' argument to :func:`crested.tl.enhancer_design_in_silico_evolution`
-
-    Parameters
-    ----------
-    optimize_func
-        Function to optimize the mutated sequence based on the original prediction.
-
-    See Also
-    --------
-    crested.tl.enhancer_design_in_silico_evolution
-    """
-
-    def __init__(self, optimize_func: Callable[..., int]) -> None:
-        """Initialize the EnhancerOptimizer class."""
-        self.optimize_func = optimize_func
-
-    def get_best(
-        self,
-        mutated_predictions: np.ndarray,
-        original_prediction: np.ndarray,
-        target: int | np.ndarray,
-        **kwargs: dict[str, Any],
-    ) -> int:
-        """
-        Get the index of the best mutated sequence based on the original prediction.
-
-        Parameters
-        ----------
-        mutated_predictions
-            The predictions of the mutated sequences.
-        original_prediction
-            The prediction of the original sequence.
-        target
-            An integer or array representing some target to optimize for.
-            For example, this can be the target class index, or an array of target values to maximize.
-            Depends on the optimization function used.
-        **kwargs
-            Additional keyword arguments to pass to the optimization function.
-        """
-        return self.optimize_func(
-            mutated_predictions, original_prediction, target, **kwargs
-        )
-
-
-def _weighted_difference(
-    mutated_predictions: np.ndarray,
-    original_prediction: np.ndarray,
-    target: int,
-    class_penalty_weights: np.ndarray | None = None,
-):
-    if len(original_prediction.shape) == 1:
-        original_prediction = original_prediction[None]
-    n_classes = original_prediction.shape[1]
-    penalty_factor = 1 / n_classes
-
-    target_increase = mutated_predictions[:, target] - original_prediction[:, target]
-    other_increases = mutated_predictions - original_prediction
-
-    other_increases[:, target] = 0
-
-    if class_penalty_weights is None:
-        class_penalty_weights = np.ones(n_classes)
-
-    score = target_increase - penalty_factor * np.sum(
-        other_increases * class_penalty_weights, axis=1
-    )
-
-    return np.argmax(score)
 
 
 def _detect_input_type(input: str | list[str] | np.array | AnnData) -> str:
@@ -328,39 +252,3 @@ def calculate_nucleotide_distribution(
     return distribution / distribution.sum(axis=-1, keepdims=True)
 
 
-def derive_intermediate_sequences(
-    enhancer_design_intermediate: dict,
-) -> list[list[str]]:
-    """
-    Derive the intermediate sequences from the enhancer design intermediate dictionary.
-
-    Parameters
-    ----------
-    enhancer_design_intermediate
-        The enhancer design intermediate dictionary.
-
-    Returns
-    -------
-    A nested list of intermediate sequences, of shape [[seq1_step1, seq1_step2, ...], [seq2_step1, seq2_step2, ...]]
-
-    See Also
-    --------
-    crested.tl.enhancer_design_in_silico_evolution
-    crested.tl.enhancer_design_motif_insertion
-    """
-    all_designed_list = []
-    for intermediate_dict in enhancer_design_intermediate:
-        current_sequence = intermediate_dict["initial_sequence"]
-        sequence_list = [current_sequence]
-        for loc, change in intermediate_dict["changes"]:
-            if loc == -1:
-                continue
-            else:
-                current_sequence = (
-                    current_sequence[:loc]
-                    + change
-                    + current_sequence[loc + len(change) :]
-                )
-                sequence_list.append(current_sequence)
-        all_designed_list.append(sequence_list)
-    return all_designed_list
